+++ utils/distortioncorrector/pointListEditImageWidget.cpp
@@ -443,7 +443,7 @@ void PointListEditImageWidgetUnited::childRepaint(QPaintEvent *event, QWidget *w
     }
 
     /* Draw additional points*/
-    for (int i = 0; i < pointList.size(); i++)
+    for (size_t i = 0; i < pointList.size(); i++)
     {
         Vector2dd &p = pointList[i];
         Vector2dd imageCoords = imageToWidgetF(p);
